Add extra data pointer to teenyat struct
During development for a system conatining multiple TeenyAT-based
agents, it was realized there was not a simple way inside a callback
to identify data outside the teenyat instance associated with it.
For example, one agent vehicle (each with a teenyat instance) tries
to drop speed by writing to throttle hardware over the bus.  This
would result in a write callback, and from the teenyat pointer alone,
the callback (shared perhaps by all agent vehicles) could not
otherwise determine which vehicle's throttle to drop.