Change "term()" to use/return a unique_ptr instead of bool
The parser methods were all being written to use bool returns to verify
the grammar rules were met.  This required additional output parameters
for the grammar functions in order for callers to have context of
successful rule matches.  With this change, we'll start returning
pointers to some data that provides the context while the address in the
pointer, whether valid or nullptr, can also be used as the prior bool
returns.