Initial complete encoding of code_2_line
This is the first commit that includes a near-complete
attempt to encode a typical instruction.  code_2_line
was chosen because its format has the most operands.
While this encodes an instruction, it does NOT yet
produce a binary file output of any kind.  Further,
the parser is never even being called after the lexer.
We really just gain a template for how all typical
code line grammars may be implemented.