Switch p_code_X_inst() functions to return token pointers.
These functions were being written to return the opcode equivalent of the
instructions used, but there were two issue developing from this that needed
resolution.

1) duplicate code was creeping in for multiple instruction grammars to
return that, for example, the ADD instruction had the ADD opcode

2) pseudo-instructions, such as DEC, if interpreted only by their actual
opcode used (eg, DEC uses the opcode for SUB) could cause ambiguity
in the parser... was it DEC or SUB?

With the original token ptr and matched grammar rule, it's trivial to
determine the appropriate opcode to use for the matched instruction.