Registers no longer case sensitive (fixes #20)
It has nothing to do with the CMP instruction in the end. A change
was made recently to drop the old tnasm code that automatically
fixed all characters in a source file to lowercase as a means to
achieve case insensitivity since the regex added later has a case
insensitive mode, already. This had an unexpected effect in that
the register handling code in the lexer was expecting all
lowercase, which was no longer there.