Introduce structures and code for random number generation
Each teenyat instance will have a PCG-XSH-RR with 64-bit state and 16-bit output
based on the algorithm described at...
https://en.wikipedia.org/wiki/Permuted_congruential_generator

Upon initialization of the first teenyat instance in a system,
a stream counter is initialized and used to set the state and increment
of that teenyat's internal random number generator.  Additionally,
the static stream counter is incremented so that each consecutive
teenyat created in that system run will likely get the same seed
but is ensured a unique increment.