Change Makfile to work with old make on macOS
It appears the Apple version of make distributed with its
xcode command line tools is 3.81, from 2006, even though
it is 2025.  Primarily, this did not yet include support
for  invocations.  The affected Makefile
sections were rewritten to be compatible with make 3.81
so TeenyAT users would not have to manually install a more
recent make just to build TeenyAT.