Set/export the SHELL variable on Windows to prevent alternate shell issues
On one test Windows computer, because an instance of sh was
in the PATH, and no SHELL was defined, make was running recipe
commands in sh (which had been installed by Git) and failing
as they were written for the command prompt or PowerShell.
Setting the SHELL variable to cmd.exe at the root level
Makefile, avoids this issue of SHELL searching.