Delay instruction now uses prescaler in the reg1
Reg1 field of DLY encodings was ignored and encoded as rZ by tnasm.
This commit extends the delayable cycles by utilizing reg1 in the
encoded instruction to scale the count provided by reg2 + immediate.
Essentially, the new method for calucating the number of cycles to
delay is reg1 * (reg2 + Imm).  If reg1 is rZ, the prescaler
multiplier 1 is used instead of 0.