Correct misleading timing comment
Unsigned target speeds can't be negative because of unsigned type,
so the comment shouldn't suggest it could be negative.  It didn't
say we were checking for negative values, but since it did not refer
to the unsigned type used, it came across as if we should have a
"<= 0" comparison in the conditional statement.