Use a boolean to determine whether a teenyat is clocked
Previously, to make a teenyat instance run as fast as possible,
clock_rate.cycles_until_calibrate was set to a negative value.
This meant the value was signed, which didn't make sense given
it's purpose by name.  This variable has been changed to an
unsigned value to stick with its implied use, and a new boolean,
is_clocked, was introduced for a more readable check.