Corrected bug in wall_time_base
Previously, the wall_time_base time was being set in the
tny_reset() function.  Since the system writer calls this function
at their own will, it is unclear how long after reseting (also
done as a part of initializing a teenyat) the system will begin
issuing clock cycles to the instance.  As such, the time used to
initialize the wall_time_base is determined at the beginning of
the first cycle after reset/initialization.