Add "unclocked" bit to the control status register
This bit now used to determined whether the instance is unclocked (a 1)
or clocked (a 0).  It also provides a mechanism for the assembly
programmer to switch mode at runtime, although the unclocked effect
may not always be what they hope for, particularly in some multiagent
systems depending on how the system writer gives cycles to the many
TeenyAT instances.