Ensure switching to clocked mode resets CPU rate control
Whenever an instance goes from unclocked to clocked modes, we must
also reset the clock_rate epoch, cycles until calibration and time
of last calibration.  Otherwise, it is likely the cycle count and
times measured will be wildly off while not being tracked in
unclocked mode.  The instance could think it's so far ahead that it
starts extending its cycles dramatically to slow the simulation
down.