Fix error with consective random streams being identical
To ensure random streams for each teenyat, the increment must be an
arbitrary odd number.  The static "stream" value was perpetually
incremented, added to the increment, proper, and then then low bit
of the increment set to ensure it is odd.  This meant when the
increment (pre-oddification) ended in a 0, the 0 would be set to a 1.
The next increment (pre-oddification) will then end in 1, but
nonetheless have that low 1 set to a 1.

To resolve, the stream value is incremented by 2 intead of 1 to ensure
each consecutive increment (pre-oddification) has at minimum 1 differing
bit in bits 1 through 15.