Capture the reg2 + immed value first on PSH
This must be done to ensure when the stack reference register (reg1)
is the same as reg2 that the value stored feel consistent with all
other circumstances the assembly programmer is used to..  Otherwise,
the reg[reg2] value will be truncated to 15 bits as all stack
reference registers are.

For example, a programmer could run these lines:
     PSH [rA], rB  ; value in rB stored at ram[rA] and rA--
     POP rB, [rA]  ; rB restored to original value
                   ; rA may have changed because of RAM-wrapping

Without this patch, though, were rB == 0x8000 and used as the
stack reference register (instead of rA) in the code above.
Say the following are the first two lines of code of a program...
     PSH [rB], rB
     POP rB, [rB]
0) before anything runs, the PSH line (at address 0x0000) is 0x1E40
1) on PSH, rB truncated to 0x0000 and stored at 0x0000
2) finishing PSH, rB is dec/wrapped to 0x7FFF as the reference stack
   pointer
3) on POP, rB's new value (0x7FFF) gets inc/wrapped to 0x0000
4) finishing POP, rB is replaced with ram[0x0000] which was set to
   0x0000 in step 1
5) yikes... PSH-ing then POP-ing rB this way would not have restored
   rB like you would expect.