Correct the computaion of the carry flag
Instructions that affect the carry flag all do it by performing some arithmetic
in 32 bits and evaluate the result's bit 16 as the carry.  The prior code was
incorrectly expanding the 16-bit values used in the arithmetic such that bits in the
higher 16 bits might not be 0 prior to the arithmetic.  This allowed what would have
been the carry bit to be manipulated in the arithmetic along with one or more 1s in
those higher bits, corrupting the bit used to evaluate the carry.